% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{everforest_palette}
\alias{everforest_palette}
\alias{dark_foreground}
\alias{dark_background}
\alias{light_foreground}
\alias{light_background}
\title{Everforest official color palette(s)}
\format{
An object of class \code{list} of length 14.

An object of class \code{list} of length 3.

An object of class \code{list} of length 14.

An object of class \code{list} of length 3.
}
\source{
https://github.com/sainnhe/everforest
}
\usage{
dark_foreground

dark_background

light_foreground

light_background
}
\description{
These are the official colors as listed on the
\href{https://github.com/sainnhe/everforest/blob/master/palette.md}{Everforest palette page}.
}
\keyword{datasets}
