% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{everforest_rstheme}
\alias{everforest_rstheme}
\title{Create Everforest .rstheme file}
\usage{
everforest_rstheme(
  dark = TRUE,
  variant = "medium",
  apply = TRUE,
  as_sass = FALSE,
  ...
)
}
\arguments{
\item{dark}{A Boolean - use dark (TRUE) or light (FALSE) theme;
defaults to TRUE}

\item{variant}{A string - the theme palette variant to use
(hard, medium, soft); see \href{https://github.com/sainnhe/everforest}{Everforest}}

\item{apply}{A Boolean - generate and install/apply the .rstheme file (TRUE),
or just save the .rstheme file without applying (FALSE)}

\item{as_sass}{A Boolean - save theme as .sass file and do not apply (TRUE),
defaults to FALSE}

\item{...}{Additional arguments to pass to \code{\link[rsthemes:rstheme]{rsthemes::rstheme()}} overriding
defaults if necessary}
}
\description{
Create Everforest .rstheme file
}
